remove erroneous trailing comma during cms:*:dump
some logic had been specifically added to ensure that a trailing comma
would not be added to the "dump" output, which outputs as an array of
multiple json objects.

However, due to a copy&paste error, the "final line" output exactly as
the rest of the lines, despite having been (logically) broken out
separately. This meant that the output was *not* valid json.

This was not noticed for a long time, as the type of post-processing
which was being done on the output of this dump rarely involved reading
the entire json array

Here we correct this, so that when we detect that we are in the final
line, the trailing comma is properly skipped.